# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Almera E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UploadResponse {
  url: string;
}

export interface ProductVariant {
  size: string;
  color: string;
  quality: string;
  stock: number;
}

export interface Product {
  id: string;
  name: string;
  slug: string;
  description: string;
  category: string;
  originalPrice: number;
  discountedPrice: number;
  images: string[];
  videos?: string[];
  variants: ProductVariant[];
  rating: number;
  reviewCount: number;
  isFeatured: boolean;
  isNew: boolean;
}

export interface ProductUpsertRequest {
  name: string;
  slug: string;
  description: string;
  category: string;
  originalPrice: number;
  discountedPrice: number;
  images?: string[];
  videos?: string[];
  variants?: ProductVariant[];
  isFeatured?: boolean;
  isNew?: boolean;
}

export interface Review {
  id: string;
  productId: string;
  productSlug: string;
  productName: string;
  userId?: string | null;
  authorName: string;
  rating: number;
  title?: string | null;
  body: string;
  createdAt: string;
}

export interface CreateReviewRequest {
  authorName: string;
  rating: number;
  title?: string;
  body: string;
}

export interface Address {
  id: string;
  label: string;
  recipientName: string;
  phone: string;
  address: string;
  city: string;
  isDefault: boolean;
}

export interface CreateAddressRequest {
  label?: string;
  recipientName: string;
  phone: string;
  address: string;
  city: string;
  isDefault?: boolean;
}

export interface AccountProfile {
  id: string;
  email: string;
  fullName?: string | null;
  phone?: string | null;
  createdAt: string;
}

export interface UpdateProfileRequest {
  fullName?: string;
  phone?: string;
  currentPassword?: string;
  newPassword?: string;
}

export interface OrderItem {
  productId: string;
  productName: string;
  productImage?: string | null;
  size: string;
  color: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export interface CreateOrderRequest {
  customerName: string;
  customerEmail?: string | null;
  customerPhone: string;
  shippingAddress: string;
  shippingCity: string;
  paymentMethod: string;
  paymentScreenshotUrl?: string | null;
  notes?: string | null;
  items: OrderItem[];
}

export interface Order {
  id: string;
  orderNumber: string;
  customerName: string;
  customerEmail?: string | null;
  customerPhone: string;
  shippingAddress: string;
  shippingCity: string;
  paymentMethod: string;
  paymentScreenshotUrl?: string | null;
  paymentStatus: string;
  status: string;
  subtotal: number;
  deliveryFee: number;
  codFee: number;
  total: number;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type OrderWithItems = Order & {
  items: OrderItem[];
};

export interface UpdateOrderRequest {
  status?: string;
  paymentStatus?: string;
}

export interface PaginatedOrders {
  data: OrderWithItems[];
  total: number;
  page: number;
  limit: number;
}

export interface DashboardStats {
  totalOrders: number;
  revenue: number;
  pendingOrders: number;
  deliveredOrders: number;
}

export interface RevenueDataPoint {
  date: string;
  revenue: number;
  orders: number;
}

export interface BestSeller {
  productId: string;
  productName: string;
  totalSold: number;
  totalRevenue: number;
}

export type ExtendedStatsStatusBreakdownItem = {
  status: string;
  count: number;
};

export type ExtendedStatsCategoryBreakdownItem = {
  category: string;
  revenue: number;
  orders: number;
};

export interface ExtendedStats {
  totalOrders: number;
  totalRevenue: number;
  totalProfit: number;
  pendingOrders: number;
  processingOrders: number;
  deliveredOrders: number;
  cancelledOrders: number;
  revenueByDay: RevenueDataPoint[];
  bestSellers: BestSeller[];
  statusBreakdown: ExtendedStatsStatusBreakdownItem[];
  categoryBreakdown: ExtendedStatsCategoryBreakdownItem[];
}

export interface LegalPage {
  id: string;
  slug: string;
  title: string;
  content: string;
  updatedAt: string;
}

export interface UpdateLegalPageRequest {
  title?: string;
  content?: string;
}

export interface SessionUser {
  id: string;
  email?: string | null;
  fullName?: string | null;
  role: string;
  isLoggedIn: boolean;
}

export interface RegisterRequest {
  email: string;
  password: string;
  fullName?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type ListProductsParams = {
  category?: string;
  featured?: boolean;
};

export type TrackOrderParams = {
  orderNumber?: string;
  phone?: string;
};

export type AdminListOrdersParams = {
  status?: string;
  page?: number;
  limit?: number;
};

export type GetSettings200 = { [key: string]: string };

export type AdminUpdateSettingsBody = { [key: string]: string };

export type AdminUploadProductImageBody = {
  file?: Blob;
};

export type UploadPaymentScreenshotBody = {
  file?: Blob;
};