# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Almera E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /account/me:
    get:
      operationId: getAccountProfile
      tags: [account]
      responses:
        "200":
          description: Account profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountProfile"
    put:
      operationId: updateAccountProfile
      tags: [account]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountProfile"
...
  /account/addresses:
    get:
      operationId: listAddresses
      tags: [account]
      responses:
        "200":
          description: User addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: createAddress
      tags: [account]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressRequest"
      responses:
        "201":
          description: Address created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /account/addresses/{id}:
    put:
      operationId: updateAddress
      tags: [account]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressRequest"
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    delete:
      operationId: deleteAddress
      tags: [account]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Address:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
        recipientName: { type: string }
        phone: { type: string }
        address: { type: string }
        city: { type: string }
        isDefault: { type: boolean }
      required: [id, label, recipientName, phone, address, city, isDefault]
    CreateAddressRequest:
      type: object
      properties:
        label: { type: string }
        recipientName: { type: string }
        phone: { type: string }
        address: { type: string }
        city: { type: string }
        isDefault: { type: boolean }
      required: [recipientName, phone, address, city]
    AccountProfile:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        fullName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        createdAt: { type: string }
      required: [id, email, createdAt]
    UpdateProfileRequest:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
        currentPassword: { type: string }
        newPassword: { type: string }
...